src/backend/requirements.txt
# src/backend/requirements.txt
Flask==2.0.1
PyYAML==6.0
gunicorn==20.1.0
python-dotenv==0.19.0
prometheus-client==0.12.0
python-json-logger==2.0.2
SQLAlchemy==1.4.27
psycopg2-binary==2.9.2